requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
prometheus-client==0.20.0
Werkzeug==3.0.1
RPi.GPIO==0.7.1
w1thermsensor~=2.3.0